// libs/ui/src/tokens/colors/color.type.ts
export interface IColorTokenConfig {
  textColor?: IColor;
  textDecorationColor?: IColor;
  bgColor?: IColor;
  borderColor?: IColor;
  divideColor?: IColor;
  outlineColor?: IColor;
  ringColor?: IColor;
  ringOffsetColor?: IColor;
  boxShadowColor?: IColor;
  accentColor?: IColor;
  caretColor?: IColor;
}

type IColor = string;